refactor(vfs-rpc-server): use UUID v4 session ids and split logic into core lib
session_id
----------
The session id is sent on every RPC request and used by the server only
for logging. The previous generator emitted a 6-character alphanumeric
hash of a per-process counter plus wall-clock nanoseconds — roughly
2^31 effective entropy, with birthday-bound collisions starting around
50k sessions. Switch to a UUID v4 in `simple` form (32 hex chars, 122
bits of entropy).

We can't use `uuid`'s built-in `v4` feature: it depends transitively on
`wasip2` 0.2.9, which the component linker cannot merge with the
0.2.6 wasi interfaces our world imports. Instead, import
`wasi:random/random@0.2.6` in `wit/rpc-server.wit`, ask for 16 bytes,
and feed them into `Uuid::Builder::from_random_bytes`. A
`#[cfg(not(target_family = "wasm"))]` fallback uses the wall clock so
host tests still work.

Crate split
-----------
The cdylib previously bundled three concerns:
  - WIT bindings + WASI TCP/socket I/O loop
  - protobuf decode / encode of each request
  - the actual request handler (filesystem operations, S3 sync glue)

Only the first concern requires `wit-bindgen` and `cdylib`. Extract the
rest into a new `vfs-rpc-server-core` library crate:

  - `ServerContext` struct + `init_server` + `map_fs_error`
  - `ServerContext::handle_request`, async, takes a pre-generated
    session id so the random-source dependency stays in the shell

Unit tests in `vfs-rpc-server-core` cover (a) every `FsError`-to-
`ErrorCode` mapping, (b) the `Connect` protocol-version check in both
directions, (c) an end-to-end open/write/seek/read round trip through
the handler, and (d) `Unlink` followed by `Stat` returning `NotFound`.
None of these required the cdylib build before. The shell keeps a
small `tests` module exercising the UUID format/uniqueness.

E2E (host-trait Tier 1 + RPC Tier 2) passes after the split.
