+++ crates/rpc/vfs-rpc-server-core/src/lib.rs
@@ -1,0 +1,636 @@
+//! Pure RPC request-handling logic for the VFS RPC server.
+//!
+//! This crate is the "logic layer" extracted from `vfs-rpc-server`, the
+//! cdylib component that runs on `wasm32-wasip2`. The cdylib depends on
+//! this crate, hands incoming protobuf-decoded `Request`s to
+//! [`ServerContext::handle_request`], and ships the resulting `Response`
+//! back over its TCP listener.
+//!
+//! Splitting these two concerns has a single concrete benefit: this crate
+//! is plain Rust (no `wit-bindgen`, no `cdylib`), so it builds and tests
+//! under `cargo test --workspace` against the native host like any other
+//! library. The cdylib stays minimal — TCP I/O loop, WASI imports and a
+//! UUID generator using `wasi:random/random` — and is exercised through
+//! the e2e suite.
+
+use std::cell::RefCell;
+use std::collections::{HashMap, HashSet};
+use std::rc::Rc;
+
+use fs_core::{Fs, FsError, MonotonicCounter};
+use vfs_rpc_protocol::{DirEntry, ErrorCode, FileMetadata, Request, Response, PROTOCOL_VERSION};
+
+#[cfg(feature = "s3-sync")]
+use std::sync::Arc;
+#[cfg(feature = "s3-sync")]
+use vfs_sync_adapter::{init_from_s3, new_s3_storage, MetadataCache, SyncConfig, SyncManager};
+
+/// Server context holding shared state. Constructed once via
+/// [`init_server`] and shared across all client handlers.
+pub struct ServerContext {
+    pub fs: Rc<RefCell<Fs<MonotonicCounter>>>,
+    #[cfg(feature = "s3-sync")]
+    pub sync_manager: Option<SyncManager<MonotonicCounter>>,
+    /// Map from file descriptor to path for sync tracking.
+    pub fd_path_map: RefCell<HashMap<u32, String>>,
+    /// Track which fds have been refreshed from S3 (to avoid repeated
+    /// refreshes within the lifetime of an fd).
+    pub s3_refreshed_fds: RefCell<HashSet<u32>>,
+    /// Whether to refresh files from S3 on read (`VFS_READ_MODE=s3`).
+    #[cfg(feature = "s3-sync")]
+    pub read_from_s3: bool,
+    /// Whether to check S3 metadata on open (`VFS_METADATA_MODE=s3`).
+    #[cfg(feature = "s3-sync")]
+    pub metadata_sync: bool,
+}
+
+impl ServerContext {
+    /// Handle a single RPC request and return the response.
+    ///
+    /// `session_id` is the client's existing session id (or `None` before
+    /// `Connect`). `new_session_id` is pre-generated by the caller and is
+    /// used **only** to populate the response for `Request::Connect`; for
+    /// all other request kinds it is ignored. The caller pre-generates so
+    /// that the source of randomness (`wasi:random` on WASI, an OS RNG on
+    /// host tests) is kept outside this pure library.
+    pub async fn handle_request(
+        &self,
+        request: Request,
+        session_id: Option<String>,
+        new_session_id: String,
+    ) -> Response {
+        if let Some(ref sid) = session_id {
+            log::debug!("[session {}] Processing request", sid);
+        }
+
+        match request {
+            Request::Connect { version } => {
+                if version != PROTOCOL_VERSION {
+                    Response::Error {
+                        code: ErrorCode::ProtocolError,
+                        message: format!(
+                            "Protocol version mismatch: client={}, server={}",
+                            version, PROTOCOL_VERSION
+                        ),
+                    }
+                } else {
+                    log::info!("[session {}] New client connected", new_session_id);
+                    Response::Connected {
+                        session_id: new_session_id,
+                        version: PROTOCOL_VERSION,
+                    }
+                }
+            }
+
+            Request::OpenPath { path, flags } => {
+                // Check S3 metadata and refresh if changed (when metadata sync is enabled).
+                #[cfg(feature = "s3-sync")]
+                if self.metadata_sync {
+                    if let Some(ref sync) = self.sync_manager {
+                        match sync.check_and_refresh_from_s3(&path).await {
+                            Ok(refreshed) => {
+                                if refreshed {
+                                    log::debug!("[sync] Refreshed on open: {}", path);
+                                }
+                            }
+                            Err(e) => {
+                                log::error!("[sync] Metadata check failed for {}: {}", path, e);
+                            }
+                        }
+                    }
+                }
+
+                match self.fs.borrow_mut().open_path_with_flags(&path, flags) {
+                    Ok(fd) => {
+                        self.fd_path_map.borrow_mut().insert(fd, path);
+                        Response::Fd { fd }
+                    }
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::Read { fd, length } => {
+                // Refresh from S3 before read (once per fd, only when read-through is enabled).
+                #[cfg(feature = "s3-sync")]
+                if self.read_from_s3 && !self.s3_refreshed_fds.borrow().contains(&fd) {
+                    let path = self.fd_path_map.borrow().get(&fd).cloned();
+                    if let Some(path) = path {
+                        if let Some(ref sync) = self.sync_manager {
+                            if let Err(e) = sync.refresh_file_from_s3(&path).await {
+                                log::error!("[sync] Refresh failed for {}: {}", path, e);
+                            }
+                        }
+                    }
+                    self.s3_refreshed_fds.borrow_mut().insert(fd);
+                }
+
+                let mut buf = vec![0u8; length];
+                match self.fs.borrow_mut().read(fd, &mut buf) {
+                    Ok(n) => {
+                        buf.truncate(n);
+                        Response::Data { bytes: buf }
+                    }
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::Write { fd, data } => {
+                let result = self.fs.borrow_mut().write(fd, &data);
+                match result {
+                    Ok(n) => {
+                        #[cfg(feature = "s3-sync")]
+                        if let Some(ref sync) = self.sync_manager {
+                            let path = self.fd_path_map.borrow().get(&fd).cloned();
+                            if let Some(path) = path {
+                                if sync.is_realtime() {
+                                    if let Err(e) = sync.sync_file_now(&path).await {
+                                        log::error!("[sync] RealTime sync failed: {}", e);
+                                    }
+                                } else {
+                                    sync.enqueue_upload(path);
+                                }
+                            }
+                        }
+                        Response::Written { count: n }
+                    }
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::Close { fd } => {
+                self.fd_path_map.borrow_mut().remove(&fd);
+                self.s3_refreshed_fds.borrow_mut().remove(&fd);
+                match self.fs.borrow_mut().close(fd) {
+                    Ok(()) => Response::Ok,
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::Seek { fd, offset, whence } => {
+                match self.fs.borrow_mut().seek(fd, offset, whence) {
+                    Ok(pos) => Response::Position { pos },
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::Ftruncate { fd, size } => match self.fs.borrow_mut().ftruncate(fd, size) {
+                Ok(()) => {
+                    #[cfg(feature = "s3-sync")]
+                    if let Some(ref sync) = self.sync_manager {
+                        if let Some(path) = self.fd_path_map.borrow().get(&fd) {
+                            sync.enqueue_upload(path.clone());
+                        }
+                    }
+                    Response::Ok
+                }
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::Fstat { fd } => match self.fs.borrow().fstat(fd) {
+                Ok(meta) => Response::Metadata {
+                    metadata: FileMetadata {
+                        size: meta.size,
+                        created: meta.created,
+                        modified: meta.modified,
+                        is_dir: meta.is_dir,
+                    },
+                },
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::Stat { path } => match self.fs.borrow().stat(&path) {
+                Ok(meta) => Response::Metadata {
+                    metadata: FileMetadata {
+                        size: meta.size,
+                        created: meta.created,
+                        modified: meta.modified,
+                        is_dir: meta.is_dir,
+                    },
+                },
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::Mkdir { path } => match self.fs.borrow_mut().mkdir(&path) {
+                Ok(()) => Response::Ok,
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::MkdirP { path } => match self.fs.borrow_mut().mkdir_p(&path) {
+                Ok(()) => Response::Ok,
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::Unlink { path } => {
+                let result = self.fs.borrow_mut().unlink(&path);
+                match result {
+                    Ok(()) => {
+                        #[cfg(feature = "s3-sync")]
+                        if let Some(ref sync) = self.sync_manager {
+                            sync.enqueue_delete(path.clone());
+                        }
+                        Response::Ok
+                    }
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::Readdir { path } => match self.fs.borrow().readdir(&path) {
+                Ok(names) => {
+                    let fs = self.fs.borrow();
+                    let mut entries = Vec::new();
+                    for name in names {
+                        let full_path = if path == "/" {
+                            format!("/{}", name)
+                        } else {
+                            format!("{}/{}", path, name)
+                        };
+                        let is_dir = fs.stat(&full_path).map(|meta| meta.is_dir).unwrap_or(false);
+                        entries.push(DirEntry { name, is_dir });
+                    }
+                    Response::DirEntries { entries }
+                }
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::ReaddirFd { fd } => match self.fs.borrow().readdir_fd(fd) {
+                Ok(entries) => {
+                    let dir_entries = entries
+                        .into_iter()
+                        .map(|(name, is_dir)| DirEntry { name, is_dir })
+                        .collect();
+                    Response::DirEntries {
+                        entries: dir_entries,
+                    }
+                }
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::Rmdir { path } => match self.fs.borrow_mut().rmdir(&path) {
+                Ok(()) => Response::Ok,
+                Err(e) => map_fs_error(e),
+            },
+
+            Request::Rename { old_path, new_path } => {
+                match self.fs.borrow_mut().rename(&old_path, &new_path) {
+                    Ok(()) => Response::Ok,
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::AppendWrite { fd, data } => {
+                let result = self.fs.borrow_mut().append_write(fd, &data);
+                match result {
+                    Ok(n) => {
+                        #[cfg(feature = "s3-sync")]
+                        if let Some(ref sync) = self.sync_manager {
+                            let path = self.fd_path_map.borrow().get(&fd).cloned();
+                            if let Some(path) = path {
+                                if sync.is_realtime() {
+                                    if let Err(e) = sync.sync_file_now(&path).await {
+                                        log::error!("[sync] RealTime sync failed: {}", e);
+                                    }
+                                } else {
+                                    sync.enqueue_upload(path);
+                                }
+                            }
+                        }
+                        Response::Written { count: n }
+                    }
+                    Err(e) => map_fs_error(e),
+                }
+            }
+
+            Request::OpenAt {
+                dir_fd,
+                path,
+                flags,
+            } => match self.fs.borrow_mut().open_at(dir_fd, &path, flags) {
+                Ok(fd) => {
+                    // Compute absolute path for sync tracking.
+                    if let Some(dir_path) = self.fd_path_map.borrow().get(&dir_fd) {
+                        let abs_path = if dir_path == "/" {
+                            format!("/{}", path.trim_start_matches('/'))
+                        } else {
+                            format!(
+                                "{}/{}",
+                                dir_path.trim_end_matches('/'),
+                                path.trim_start_matches('/')
+                            )
+                        };
+                        self.fd_path_map.borrow_mut().insert(fd, abs_path);
+                    }
+                    Response::Fd { fd }
+                }
+                Err(e) => map_fs_error(e),
+            },
+        }
+    }
+}
+
+/// Map a `fs-core` error to the protocol's error response.
+pub fn map_fs_error(error: FsError) -> Response {
+    let (code, message) = match error {
+        FsError::NotFound => (ErrorCode::NotFound, "Not found"),
+        FsError::NotADirectory => (ErrorCode::NotADirectory, "Not a directory"),
+        FsError::IsADirectory => (ErrorCode::IsADirectory, "Is a directory"),
+        FsError::InvalidArgument => (ErrorCode::InvalidArgument, "Invalid argument"),
+        FsError::BadFileDescriptor => (ErrorCode::BadFileDescriptor, "Bad file descriptor"),
+        FsError::PermissionDenied => (ErrorCode::PermissionDenied, "Permission denied"),
+        FsError::AlreadyExists => (ErrorCode::AlreadyExists, "Already exists"),
+        FsError::NotEmpty => (ErrorCode::NotEmpty, "Directory not empty"),
+    };
+
+    Response::Error {
+        code,
+        message: message.to_string(),
+    }
+}
+
+/// Build a fresh [`ServerContext`], optionally pre-populating it from an
+/// S3 bucket when the `s3-sync` feature is enabled and `VFS_S3_BUCKET` is
+/// set. Reads configuration entirely from environment variables.
+pub async fn init_server() -> ServerContext {
+    #[cfg(feature = "s3-sync")]
+    let s3_bucket = std::env::var("VFS_S3_BUCKET").ok();
+    #[cfg(feature = "s3-sync")]
+    let s3_prefix = std::env::var("VFS_S3_PREFIX").unwrap_or_else(|_| "vfs/".to_string());
+
+    #[cfg(feature = "s3-sync")]
+    let read_from_s3 = std::env::var("VFS_READ_MODE")
+        .map(|v| v.to_lowercase() == "s3")
+        .unwrap_or(false);
+    #[cfg(feature = "s3-sync")]
+    let metadata_sync = std::env::var("VFS_METADATA_MODE")
+        .map(|v| v.to_lowercase() == "s3")
+        .unwrap_or(false);
+
+    #[cfg(feature = "s3-sync")]
+    if read_from_s3 {
+        log::info!("Read-through mode enabled (VFS_READ_MODE=s3)");
+    }
+    #[cfg(feature = "s3-sync")]
+    if metadata_sync {
+        log::info!("Metadata sync mode enabled (VFS_METADATA_MODE=s3)");
+    }
+
+    #[cfg(feature = "s3-sync")]
+    let (fs, sync_manager) = if let Some(bucket) = s3_bucket {
+        log::info!(
+            "S3 persistence enabled: bucket={}, prefix={}",
+            bucket,
+            s3_prefix
+        );
+
+        let s3 = Arc::new(new_s3_storage(bucket, s3_prefix).await);
+
+        let (fs, metadata_cache) = match init_from_s3::<MonotonicCounter>(&s3).await {
+            Ok((fs, cache)) => (fs, cache),
+            Err(e) => {
+                log::error!(
+                    "Failed to load from S3: {}, starting with empty filesystem",
+                    e
+                );
+                (Rc::new(RefCell::new(Fs::new())), MetadataCache::new())
+            }
+        };
+
+        let config = SyncConfig::default();
+        log::info!(
+            "Sync mode: {:?} (set VFS_SYNC_MODE=realtime for immediate sync)",
+            config.mode
+        );
+        let sync_manager = SyncManager::new(
+            s3,
+            vfs_sync_adapter::AdapterFs(fs.clone()),
+            metadata_cache,
+            config,
+        );
+
+        (fs, Some(sync_manager))
+    } else {
+        log::info!("S3 persistence disabled (VFS_S3_BUCKET not set)");
+        (Rc::new(RefCell::new(Fs::new())), None)
+    };
+
+    #[cfg(not(feature = "s3-sync"))]
+    let fs = Rc::new(RefCell::new(Fs::new()));
+
+    ServerContext {
+        fs,
+        #[cfg(feature = "s3-sync")]
+        sync_manager,
+        fd_path_map: RefCell::new(HashMap::new()),
+        s3_refreshed_fds: RefCell::new(HashSet::new()),
+        #[cfg(feature = "s3-sync")]
+        read_from_s3,
+        #[cfg(feature = "s3-sync")]
+        metadata_sync,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn block_on<F: std::future::Future>(f: F) -> F::Output {
+        // Minimal single-thread tokio runtime for tests; the production
+        // cdylib also uses `new_current_thread`.
+        tokio::runtime::Builder::new_current_thread()
+            .enable_all()
+            .build()
+            .unwrap()
+            .block_on(f)
+    }
+
+    fn empty_ctx() -> ServerContext {
+        ServerContext {
+            fs: Rc::new(RefCell::new(Fs::new())),
+            #[cfg(feature = "s3-sync")]
+            sync_manager: None,
+            fd_path_map: RefCell::new(HashMap::new()),
+            s3_refreshed_fds: RefCell::new(HashSet::new()),
+            #[cfg(feature = "s3-sync")]
+            read_from_s3: false,
+            #[cfg(feature = "s3-sync")]
+            metadata_sync: false,
+        }
+    }
+
+    #[test]
+    fn map_fs_error_translates_each_variant() {
+        let cases = [
+            (FsError::NotFound, ErrorCode::NotFound, "Not found"),
+            (
+                FsError::NotADirectory,
+                ErrorCode::NotADirectory,
+                "Not a directory",
+            ),
+            (
+                FsError::IsADirectory,
+                ErrorCode::IsADirectory,
+                "Is a directory",
+            ),
+            (
+                FsError::InvalidArgument,
+                ErrorCode::InvalidArgument,
+                "Invalid argument",
+            ),
+            (
+                FsError::BadFileDescriptor,
+                ErrorCode::BadFileDescriptor,
+                "Bad file descriptor",
+            ),
+            (
+                FsError::PermissionDenied,
+                ErrorCode::PermissionDenied,
+                "Permission denied",
+            ),
+            (
+                FsError::AlreadyExists,
+                ErrorCode::AlreadyExists,
+                "Already exists",
+            ),
+            (
+                FsError::NotEmpty,
+                ErrorCode::NotEmpty,
+                "Directory not empty",
+            ),
+        ];
+        for (err, expected_code, expected_message) in cases {
+            match map_fs_error(err) {
+                Response::Error { code, message } => {
+                    assert_eq!(code, expected_code);
+                    assert_eq!(message, expected_message);
+                }
+                _ => panic!("expected Response::Error for {:?}", err),
+            }
+        }
+    }
+
+    #[test]
+    fn connect_with_matching_version_returns_session_id() {
+        let ctx = empty_ctx();
+        let req = Request::Connect {
+            version: PROTOCOL_VERSION,
+        };
+        let resp = block_on(ctx.handle_request(req, None, "test-session".into()));
+        match resp {
+            Response::Connected {
+                session_id,
+                version,
+            } => {
+                assert_eq!(session_id, "test-session");
+                assert_eq!(version, PROTOCOL_VERSION);
+            }
+            other => panic!("expected Connected, got {:?}", other),
+        }
+    }
+
+    #[test]
+    fn connect_with_wrong_version_returns_protocol_error() {
+        let ctx = empty_ctx();
+        let req = Request::Connect {
+            version: PROTOCOL_VERSION + 1,
+        };
+        let resp = block_on(ctx.handle_request(req, None, "ignored".into()));
+        match resp {
+            Response::Error { code, .. } => {
+                assert_eq!(code, ErrorCode::ProtocolError);
+            }
+            other => panic!("expected Error, got {:?}", other),
+        }
+    }
+
+    #[test]
+    fn open_then_write_then_read_round_trips_via_handler() {
+        let ctx = empty_ctx();
+        let s = || "irrelevant".to_string();
+
+        // Open /hello.txt with O_CREAT | O_RDWR
+        let resp = block_on(ctx.handle_request(
+            Request::OpenPath {
+                path: "/hello.txt".into(),
+                flags: fs_core::O_CREAT | fs_core::O_RDWR,
+            },
+            None,
+            s(),
+        ));
+        let fd = match resp {
+            Response::Fd { fd } => fd,
+            other => panic!("expected Fd, got {:?}", other),
+        };
+
+        // Write
+        let resp = block_on(ctx.handle_request(
+            Request::Write {
+                fd,
+                data: b"hi".to_vec(),
+            },
+            None,
+            s(),
+        ));
+        assert!(matches!(resp, Response::Written { count: 2 }));
+
+        // Seek to 0
+        let _ = block_on(ctx.handle_request(
+            Request::Seek {
+                fd,
+                offset: 0,
+                whence: 0,
+            },
+            None,
+            s(),
+        ));
+
+        // Read back
+        let resp = block_on(ctx.handle_request(Request::Read { fd, length: 16 }, None, s()));
+        match resp {
+            Response::Data { bytes } => assert_eq!(bytes, b"hi"),
+            other => panic!("expected Data, got {:?}", other),
+        }
+    }
+
+    #[test]
+    fn read_on_unknown_fd_returns_bad_file_descriptor() {
+        let ctx = empty_ctx();
+        let resp =
+            block_on(ctx.handle_request(Request::Read { fd: 42, length: 8 }, None, "x".into()));
+        match resp {
+            Response::Error { code, .. } => {
+                assert_eq!(code, ErrorCode::BadFileDescriptor);
+            }
+            other => panic!("expected Error, got {:?}", other),
+        }
+    }
+
+    #[test]
+    fn unlink_then_stat_returns_not_found() {
+        let ctx = empty_ctx();
+        let s = || "x".to_string();
+
+        // Create a file then close it
+        let fd_resp = block_on(ctx.handle_request(
+            Request::OpenPath {
+                path: "/x".into(),
+                flags: fs_core::O_CREAT | fs_core::O_WRONLY,
+            },
+            None,
+            s(),
+        ));
+        let fd = match fd_resp {
+            Response::Fd { fd } => fd,
+            other => panic!("expected Fd, got {:?}", other),
+        };
+        let _ = block_on(ctx.handle_request(Request::Close { fd }, None, s()));
+
+        // Unlink and verify Stat now returns NotFound.
+        let r = block_on(ctx.handle_request(Request::Unlink { path: "/x".into() }, None, s()));
+        assert!(matches!(r, Response::Ok));
+
+        let r = block_on(ctx.handle_request(Request::Stat { path: "/x".into() }, None, s()));
+        match r {
+            Response::Error { code, .. } => assert_eq!(code, ErrorCode::NotFound),
+            other => panic!("expected Error(NotFound), got {:?}", other),
+        }
+    }
+}
